Moved 'strings.hpp' into 'core'
