+++ src/shaka_scheme/system/core/strings.hpp
